StateSpaceTypes & Attributes refactoring
