+++ core/src/main/scala/ferjorosa/sbn/core/data/attributes/StateSpaceTypes.scala
@@ -1,5 +1,4 @@
-package ferjorosa.sbn.core.variables
-
+package ferjorosa.sbn.core.data.attributes
 
 /**
   * @constructor
